Feat: Add ball bounce
When colliding with environment (See defined collision masks in project settings) or paddles the ball bounces off of it.
Bouncing off of a paddle gives a different angle to the ball depending on where it lands on it.
Boundig off of anything else bounces with a predetermined angle (ie it doesn't vary just like on the paddle according to its position)